requirements.txt
# FastAPI 및 CORS 설정
fastapi==0.115.12
starlette==0.46.2

# 비동기 통신
aiohttp==3.12.9

# 이미지 처리
opencv-python==4.10.0.84
pillow==11.1.0

# 딥러닝
torch==2.6.0
torchvision==0.21.0

# YOLOv5, YOLOv8
ultralytics==8.3.148

# 벡터/배열 처리
numpy==1.26.4

# 기타
uvicorn==0.34.3
